DNM: podvm-builder: Use a fake go.mod file to pre-fetch the cloud-api-adaptor sources
In a hermetic environment the "git clone" line in our Dockerfile.podvm-builder file is
failing.
Rather than getting the sources at build time, we need to pre-fetch them.

Using a go.mod file that contains only the reference to the cloud-api-adaptor sources,
we can tell Hermeto to pull the dependencies as if they were actual code we need for
building a binary in our image.
Once the code is available in Hermeto's cache, we can copy it from there and put it
in the podvm-builder image.

Signed-off-by: Julien Ropé <jrope@redhat.com>