test-fbc: changelog.sh simplification
We can remove several lines from the beginning of the script, because
the test-fbc files are in the same repo than the bundle.
Then if we have a commit range for the catalog changes (by default: between the
commit for the last catalog build and HEAD), then we can directly look for the
commits in this range that affect the CSV for the bundle.
No need to look up the bundle images and find the commits that generated them.

Signed-off-by: Julien Ropé <jrope@redhat.com>